Importing the initial state of the code - this code mostly serves as a proof of concept.
As it currently stands, this code will generate a CPTableView that will
list the display_name|name, version, catalogs, installer_item_size,
uninstallable, and auto remove values from each package in the 'all'
catalog.

It's functional but in no way is it ready for primetime. I still need
to at least add a source list along the left side of the screen that
will list the available catalogs. It might even be a good idea to nest
the packages under the catalogs in that new list, then reserve the rest
of the screen space for the information available about the package.