Readonly stuff might actually be done.
I'm still not sure how much more detail should be available in the
package details pane, but it looks like enough at the moment. Obviously
there's now the conditional_items key introduced with Munki 0.8.1 that
I'm still not taking into account. It looks friction sweet though. I
think it's going to be fun building the UI components to support it:

http://code.google.com/p/munki/wiki/ConditionalItems