Changed the XIB drastically.
It seems that the best way to draw an interface around Munki
repositories is to focus on packages  since that's there the most
functionality lies. Granted, I still need to study Catalogs and
Manifests deeper but I feel that the focus on Packages is a fairly
solid center piece for the application.