Removed Adobe Options and cleaned up labels.
The Adobe Options tab and the corresponding MFPackage category have
been removed from the project. The labels for each field within the
project have been updated to be a little more in line with what an
existing Munki administrator would expect (e.g. the keys you'd find in
the plist files)