Moved the packagesViewController initialization to a better location.
Packages.xib is kind of big and a bit go heavy lifting for the browser
in terms of DOM work, so ideally I'd like to make this even more
efficient by having it construct itself completely in the background.
Right now, it still builds itself before the application is drawn on
the screen, but performance is marginally improved over having it in
AppController's awakeFromCib method.