Adding in a rough idea of how packages will be placed in the manifests.
This is just a rough idea, but it makes sense to me at the moment to
have all unused packages in the collection view at the top and allow
drag and drop to the appropriate collection view below. The
package/update can will only exist in one place at a time, meaning it
will be removed from the top collection view when it is dragged into
one of the other collection views.

Since packages can be marked as updates to other packages, it made
sense to add the segmented control beside the unused installers search
bar to allow you to filter the number of packages in that view. In
addition, drag and drop should allow the application to prevent an
update from being dropped into one of the installs or uninstalls
collections due to the nature of dragging proxies and drop zones.