Adding the AbstractModel class as well as the Manifests controller and model classes.
All model classes now have the ability to recursively scan a directory
for files via the AbstractModel class. They now output a JSON structure
that represents an array of dictionaries, each dictionary contains two
keys, "path" and "file". "path" should always end with a "/", so it
should be possible to concatenate the contents of the two keys to form
the full path on the filesystem.