Changed AbstractModel::recursivelyScanDirectory to recursivelyScanDirectory_relativePaths
Accordingly changed the logic so that the default behavior is to return
paths that are relative to the munki repo instead of exposing the full
server-side filesystem path. This should be enough context for the
client and server to communicate about a given file and be sure they
are talking about the same one.