Modified the AbstractModel (in AbstractModel.php) yet again.
This modification makes the paths returned relative no matter what, and
the resulting json structure is a dictionary of arrays where each
dictionary contains an array of a directory's contents. The array will
contain either strings, which represent file names, or other
dictionaries which represent directories. This should make things a
little easier to handle in the client-app.