Added itemNamespace, leafItemsAsNormalizedArray and description methods to the MFTreeModel
itemNamespace returns something like "manifests/Core" for the
MFTreeModel representing the "Core" item which is nested under the
"manifests" item.

leafItemsAsNormalizedArray simply trims all of the leafs from the data
structure and slaps them in a bucket for you. I'm planning to use it to
display a list of packages where I don't need to display where those
packages are located in any kind of hierarchy.

description just returns the value of [self itemName] to make things a
little more readable during debugging.