Removed a few class files in favor of MFOutlineDataSource
MFOutlineDataSource is an MFTreeModel controller that adheres to the
CPOutlineViewDataSource protocol. The code used to display any data in
an outline view can now be done very easily. For an example, look at
MFManifestsOutlineDataSource.j There are only 10 lines of code
currently in place to get the outline view to display a potentially
very complex data hierarchy.