[server-app] Added AbstractController.php and modified AbstractModel
AbstractController is now the base class for the other controller
classes. It adds the HTTPRequest method to all controllers for easier
access to the REQUEST parameters.

Also modified AbstractModel to put the burden of building the full path
to the directory which will be used by the model on this class instead
of all the other model classes. The helper methods are:

public fullPathToModelRepo()
protected buildFullPathToModelRepo($aPath)

Calling buildFullPathToModelRepo will append $aPath to the value
returned by the munkiDir() method of the same class. It also handles
adding "/" where needed.