Lots of intermediate changes
Adding the MFOutlineSelectedObject and MFManifest classes which will
represent the currently selected manifest in the CPOutlineView.

Updated MFManifestsOutlineDataSource and MFPkgsInfoDataSource classes
such that they now become the delegate for the CPOutlineView when they
are active. Currently they are logging the selected item's
itemNamespace value using console.log, so very broken for cross-browser
purposes at the moment.