Moved some view logic to MFApplicationViewController.
The control has been moved from AppController to
MFApplicationController. The sheet window as well as the Packages and
Manifests view controllers are now displayed according to the logic in
MFApplicationViewController.

This cleans up the AppController a bit to make it a little easier to
programmatically select the appropriate view based on the arguments in
the hash of the URL. Right now, it does select the appropriate view
controller, but the logic to select the appropriate MFTreeModel
instance in the outline view belongs elsewhere.