Adding in the new MFNetworkDataSource base class
It's really just there because there are multiple locations where I
need to grab JSON data from the server. Therefore I've move the logic
that takes care of that into this class and modified the other classes
(and will do so for future classes) to inherit from this
MFNetworkDataSource class.