Adding support for issue 14
The server now responds to the ?controller=settings arguments. The
MFServerSettings singleton class utilizes this and makes those server
settings available to the client-app.

This means that in order to configure the client app, you only have one
key to edit in Info.plist. Everything else should be edited in
server-app/Settings.plist