Added a drop down sheet to MFNetworkDataSource to indicate that a request for data is taking longer than expected.
From the updated API docs:

This class is able to observe the length of time that overlapping
requests are taking. For example, if there are two instances of this
class (or subclass) and they both fire off a request, this class will
count the number of milliseconds between the first request and the last
response. If that amount of time surpasses
MFNetworkDataSource_SECONDS_ALLOWED_BEFORE_SHEET, it will display a
sheet window that lets the user know that the application is still
waiting on data. Once the last response is received, the sheet will
automatically be dismissed.