Search is working!
There is no CPTreeController class as of yet, but I think this will do
in the meantime. Typing a search into the search box will flatten the
tree and display something more along the lines of a result list. I'm
not sure how this will fair on large data sets though.