Patch for issue #19 - filtering out nonplus files
Noel asked that a key or set of keys be added to the settings files
that would allow server-app to ignore (either inclusively or
exclusively) files that the admin doesn't care about. In particular,
there were .html files in his repo that MunkiFace was displaying.

The added code attempts to quickly parse each file that's encountered
during a recursive directory scan and ignores files that result in an
empty dictionary object.