Added additional filtering and a parse error indicator.
Settings.plist now supports two additional keys. "plist_extension",
which is a string that when non-empty will instruct server-app to parse
only files with that extension. The other is "exclude_extensions" which
is an array of strings that server-app will skip when encountered.

If server-app encounters a file that passes both of the above filters,
but it cannot parse that file as a plist, it will add "(PARSE_ERROR) "
to the beginning of the file name. Since the search works pretty well,
administrators can search for this prefix to see if there are any
manifests or pkgsinfo files that have some kind of syntax error. Sadly,
it won't tell you what the error is, but at least it will bring it to
your attention.