Adding the installcheck.php file
Running this script in a browser will check that the munki repo is a
valid path and is a readable directory. It also checks for a 301 or 200
status code from the SoftwareRepoURL (the munki url). It will display
the filtering that MunkiFace will use when scanning directories
according to the excluded_extensions and plist_extension keys.

Hopefully this will be the starting point for a very simple web-based
server-app configuration/installation tool.