Adding SCAuth into the mix.
SCAuth hasn't been configured in any way as of this commit, but in
theory, if a .htaccess file is placed in server-app that requires
authentication, it should force the login window to appear after
client-app makes its first request.

Again, no configuration or testing of SCAuth has been done at this
point, so if it works for you, awesome, if it doesn't, just hang in
there.