Setting up some framework for various authentication methods.
In Settings.plist, AllowAll will essentially make server-app run as
'guest'. "WebServer" will tell server-app to depend upon the server to
handle authentication and will rely on PHP_AUTH_USER and PHP_AUTH_PW to
determine if a session is active and what the current username for that
session might be. The ldap url is a copy and paste from
http://httpd.apache.org/docs/2.0/mod/mod_auth_ldap.html#authldapurl and
will cause server-app to be the authentication authority (sending out
appropriate headers, handling sessions, etc) using the authentication
url provided.

More work will need to be done to determine if a single LDAPAuthDriver
can be created to detect the appropriate settings for openldap,
OpenDirectory and ActiveDirectory or if subclasses will need to be
created to handle each type of server appropriately.