Added a check in MFNetworkDataSource that prevents the "waiting for server" sheet window from appearing too soon.
"Too soon" means if the user needs to authenticate, we don't want the
sheet preventing the user from entering their credentials when prompted
by SCAuth. We also don't want it to appear before MFServerSettings has
loaded its data. Otherwise we won't know if the server is going to
require authentication or not.