Adding some preliminary code to provide the basis for KVO.
The MVManifestObserver class is currently observing changes only in the
admin notes section of the manifest UI. More work needs to be done to
make the observation a little more robust. An issue that it currently
has is that it will trigger an observation event when the manifest is
reloaded, which is obviously not when we want to send data back to the
server.