Modifying the behavior of the authentication drivers.
This is part one of two (hopefully). In this commit, only the AllowAll
driver will allow you to interact with MunkiFace. The reason is that
server-app has changed its authentication method to be strictly session
based. The POST variables 'u' and 'p' need to be present in order to
attempt authentication, and the GET variable 'logout' will terminate a
session.

Part two will be the client-app's ability to properly handle the new
authentication method. Right now, it is able to see that authentication
is required, but it simply displays an alert instead of a login window.
Therefore, part two will be a proper login window.