Part 2 of 2 for authentication
The login window is now present and actively being displayed when an
authentication exception (conveniently exception #401) is caught. It
will create a valid session as long as the proper username and password
are provided.

There is _no_ error checking at the moment, which means the expected
behavior for an invalid username or password is currently undefined.
There are also a few usability tweaks that need to be made to the login
window to make it feel a little more like the login window in Lion. The
fact that it looks like Lion's login window wasn't intentional, but
since it does, we'd might as well make it feel like it as well.