Modified MFLoginWindow - it is now making itself the class delegate for CPURLConnection and is queuing, then restarting connections after authentication is granted.
No graphical improvements have been made yet. It still needs to give
some kind of feed back to let the user know that the authentication
attempt is being made…probably a spinner graphic or something.

Also, the alert that let's the user know the authentication attempt
failed needs to be replaced with a more elegant graphical solution.