Comet exploration
This is an all-code (no xib) exploration into near realtime data
transfer between server and client. It's no where near as complete as
the project in client-app/ and server-app/, but it's already showing a
lot of promise. To test, edit MunkiFace-Comet/Resources/Settings.plist,
then load  MunkiFace-Comet/index.html in your browser and chose one of
the options from the segmented control (manifests, catalogs, pkgsinfo).
If you clicked PkgsInfo, go into your pkgsinfo folder and create an
empty file. You should see it appear in the interface almost instantly
(~10ms max delay). If you delete a file, it will be removed from the
list just as quickly. File modifications are also being sent to the
browser, which should open a lot of doors and make MunkiFace a far more
powerful tool than it might have been otherwise.

I'm currently working on getting a CPPredicateEditor in place for an
advanced search (that's what the little gear beside the search field is
all about). Once that's done, it should make MunkiFace capable of
searching through your manifests, catalogs, and pkgsinfo files on _any_
key supported by Munki.

Another major difference is that the MunkiFace-Comet folder contains a
server/ folder. I like this because it makes the app completely
self-contained and extremely easy to deploy.

More to come…