Adding the search predicate editor.
This is really good news. It not only drastically increases the
searching capability of MunkiFace, it also serves as a proof of concept
that the CPPredicateEditor should be able to be used to build an
interface for the conditional_items manifest keys…which of course,
would rule (get it? 'rule' - it's a 'rule' editor?…bah)