Major performance improvements in MFTreeModel and MFPlistCollection
Thanks to @gregneagle for providing his rather large munki repository
to me a few months back. Without it I wouldn't have realized the
slowdown. The code was taking around 14 seconds to parse all of the
files in that example repo, but is now taking an average of 1.3
seconds. That's an performance improvement of  1,076.92% - only in the
early stages of a project can a claim like that be made.