Added a search button to the predicate editor.
The search button is able to send the predicate from the predicate
editor to the MFOutlineViewController, but the field names all need
prefixed with `representedObject.`. That's obviously not going to work
very well for usability, so I'll need to figure out how to map or mask
those field names to mnemonics.