Adding part 1 of 2 for pkgsinfo drag and drop
The outline view is now able to reogranize itself based on the user
dragging and dropping the pkgsinfo items or folders into other
locations within the structure. Part 2 of this functionality will be
informing the server of the operations and allowing it to make the
proper modifications to the filesystem.