Made MFManifestViewController a little more robust when changing contentViewControllers.
It wasn't switching out views properly and was causing a lot of
resizing and memory issues. Now it's smart enough to avoid an
instruction to add a controller if that same controller is already set
as the contentViewController, and it's able to size the new view
appropriately. For good measure, it also releases the old
contentViewController before adding the new one.